ui: discover latency in very noisy environments
We are attempting to calculate the latency between A to B by finding a flow
which has been seen in both points. Due to current limitions in Gremlin we
are not able to do so in a single query forcing us to:
- 1st find a flow which intersects with A (using a small queue)
- 2nd check if the discovered flow intersects with B
If the system is wieghted down with many unique flows running from A to C we
may fill the queue in phase 1 with flows from which we can't deduce latency
(between A to  B). Such as a case in which we had HTTP traffic (A -> B) and
many many flows of DNS traffic (A -> C).

The new implementation uses a very large queue for 1st phase but then reduces
it down to a limitted set of unique flows: comparing ANodeTID, BNodeTID,
Application. As 2nd step is expenssive we only pass on a limitted list of 10
unique flows. So from the example above we would still see the HTTP traffic (A
-> B) becuase the DNS traffic (A -> C) would only use one spot in the list of
flows passed to the 2nd phase.