openflow: Support for group tables
Support in the back-end of group tables through a mostly undocumented
Openflow 1.4 feature of openvswitch.
Requires Openflow1.5 support in the switch as using it with OF1.4 may
crash the switch.
Functional tests are provided and the testing framework has been
modified to make tests more reliables (parallelism, removal of bridge of
former failed tests).